+++ src/main/scala/Castles.scala
@@ -1,5 +1,7 @@
 package chess
 
+import cats.syntax.all.*
+
 import bitboard.OpaqueBitboard
 import bitboard.Bitboard
 import bitboard.Bitboard.given
@@ -8,8 +10,6 @@ import Pos.*
 
 opaque type Castles = Long
 object Castles extends OpaqueBitboard[Castles]:
-
-  import cats.syntax.all.*
 
   extension (c: Castles)
 
@@ -65,11 +65,10 @@ object Castles extends OpaqueBitboard[Castles]:
       blackKingSide: Boolean,
       blackQueenSide: Boolean
   ): Castles =
-    val whiteKing  = whiteKingSide.whiteKing
-    val whiteQueen = whiteQueenSide.whiteQueen
-    val blackKing  = blackKingSide.blackKing
-    val blackQueen = blackQueenSide.blackQueen
-    whiteKing | whiteQueen | blackKing | blackQueen
+    whiteKingSide.whiteKing |
+      whiteQueenSide.whiteQueen |
+      blackKingSide.blackKing |
+      blackQueenSide.blackQueen
 
   def apply(str: String): Castles = str match
     case "-" => empty
@@ -95,35 +94,3 @@ object Castles extends OpaqueBitboard[Castles]:
       (castles & castles(color, side).bb).nonEmpty
 
     def any = on(KingSide) || on(QueenSide)
-
-opaque type UnmovedRooks = Long
-object UnmovedRooks extends OpaqueBitboard[UnmovedRooks]:
-  // for lila testing only
-  val default: UnmovedRooks = UnmovedRooks(Bitboard.rank(Rank.First) | Bitboard.rank(Rank.Eighth))
-  val corners: UnmovedRooks = CORNERS
-  val none: UnmovedRooks    = empty
-
-  def apply(b: Bitboard): UnmovedRooks   = b.value
-  def apply(set: Set[Pos]): UnmovedRooks = set.foldLeft(empty)((b, p) => b | p.bb)
-
-  extension (ur: UnmovedRooks)
-    def toList: List[Pos]        = ur.occupiedSquares
-    def apply(pos: Pos): Boolean = (ur & pos.bb).nonEmpty
-
-    def without(color: Color): UnmovedRooks =
-      ur & Bitboard.rank(color.lastRank)
-
-    // Try to guess the side of the rook at postion `pos`
-    // If the position is not a ummovedRook return None
-    // If the position is a ummovedRook but there is no other rook on the
-    // same rank return Some(None) (because we cannot guess)
-    // If there are two rooks on the same rank, return the side of the rook
-    def side(pos: Pos): Option[Option[Side]] =
-      val bitboard = pos.bb
-      if (ur & bitboard).isEmpty then None
-      else
-        (ur & ~bitboard & Bitboard.rank(pos.rank)).first match
-          case Some(otherRook) =>
-            if (otherRook.file > pos.file) then Some(Some(QueenSide))
-            else Some(Some(KingSide))
-          case None => Some(None)
